+++ demo/browser.js
@@ -1,10 +1,10 @@
 'use strict';
 const ipc = require('ipc');
 
-ipc.on('shortcut-pressed', (shortcut) => {
+ipc.on('shortcut-pressed', shortcut => {
 	document.querySelector('#shortcut-title').textContent = shortcut;
 });
 
-ipc.on('shortcut-status', (status) => {
+ipc.on('shortcut-status', status => {
 	document.querySelector('#shortcut-status').textContent = status;
 });