+++ demo/shortcuts-cb-handler.js
@@ -1,0 +1,10 @@
+'use strict';
+
+module.exports = {
+	'Command+4': {
+		event: 'Command+4'
+	},
+	'Command+5': {
+		event: 'Command+5'
+	}
+};